% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimallambda.R
\name{calculate_optimal_lambda}
\alias{calculate_optimal_lambda}
\title{Shared function for calculating optimal lambda}
\usage{
calculate_optimal_lambda(
  response,
  predictors,
  lambda_range = c(-2, 2),
  resolution = 100
)
}
\arguments{
\item{response}{Response variable from object for the fitted model}

\item{predictors}{predictors from the model}

\item{lambda_range}{A numeric vector of length 2 specifying the range of lambda values to consider (default: -2 to 2).}

\item{resolution}{The number of lambda values to evaluate within the range (default: 100).}
}
\value{
A numeric value representing the optimal lambda.
}
\description{
Shared function for calculating optimal lambda
}
