% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_sim.R
\name{diagnostic_sim}
\alias{diagnostic_sim}
\title{Summary of Box Cox Diagnostics}
\usage{
diagnostic_sim(x, lower_lambda = -2, upper_lambda = 2, inc = 1)
}
\arguments{
\item{x}{A numeric vector}

\item{lower_lambda}{a value for the lower range of lambda}

\item{upper_lambda}{a value for the upper range of lambda}

\item{inc}{a value for incrementing the range of lambda for each output across the range

Errors will be omitted in the following cases:
If all values of x are not positive
If all values of x are not numeric
If the range of lambda values are not within \code{[-5, 5]}.,. An optimal value will be in this range
If increment value is less than or equal to 0
If input vector contains any NA values}
}
\value{
A data frame containing a summary of all diagnostics from each lambda value
}
\description{
This function takes in a numeric vector, a range of lambda values and a value to increment the range
The Box-Cox transformation is applied to the data and a diagnostic summary is outputted.
The summary includes: Lambda value
Skewness value
Kurtosis value (tailedness)
Shapiro Wilk P-value (test for normality)
For each value of lambda
}
\examples{
x<-c(1,2,3,4,10,15,25)
diagnostic_sim(x,lower_lambda=-2,upper_lambda=2,inc= 0.5)


}
\author{
Killian Slater
}
