% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimallambda.R
\name{estimate_optimal_lambda.data.frame}
\alias{estimate_optimal_lambda.data.frame}
\title{objects of class data.frame}
\usage{
\method{estimate_optimal_lambda}{data.frame}(object, lambda_range = c(-2, 2), resolution = 100)
}
\arguments{
\item{object}{An object (data.frame or matrix) containing the response variable and predictors.}

\item{lambda_range}{A numeric vector of length 2 specifying the range of lambda values to consider (default: -2 to 2).}

\item{resolution}{The number of lambda values to evaluate within the range (default: 100).}
}
\value{
A numeric value representing the optimal lambda.
}
\description{
objects of class data.frame
}
\examples{
set.seed(123)
df <- data.frame(
  y = rexp(100, rate = 0.2),
  x1 = runif(100, 1, 10),
  x2 = runif(100, 5, 15)
 )
 lambda_optimal_df <- estimate_optimal_lambda.data.frame(df, lambda_range = c(-2, 2))
 print(lambda_optimal_df)
}
