% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimallambda.R
\name{estimate_optimal_lambda.default}
\alias{estimate_optimal_lambda.default}
\title{default method}
\usage{
\method{estimate_optimal_lambda}{default}(object, ...)
}
\arguments{
\item{object}{an object for the default lambda estimation}

\item{...}{additional parameters}
}
\description{
default method
}
