% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimallambda.R
\name{estimate_optimal_lambda.numeric}
\alias{estimate_optimal_lambda.numeric}
\title{Method for objects of class numeric}
\usage{
\method{estimate_optimal_lambda}{numeric}(object, lambda_range = c(-2, 2), resolution = 100)
}
\arguments{
\item{object}{Object of type numeric vector}

\item{lambda_range}{A numeric vector of length 2 specifying the range of lambda values to consider (default: -2 to 2).}

\item{resolution}{The number of lambda values to evaluate within the range (default: 100).}
}
\value{
A numeric value representing the optimal lambda.
}
\description{
Method for objects of class numeric
}
