% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverseboxcox.R
\name{inverse_boxcox}
\alias{inverse_boxcox}
\title{Inverse Box-Cox Transformation}
\usage{
inverse_boxcox(x, lambda)
}
\arguments{
\item{x}{Transformed data (numeric vector or matrix).}

\item{lambda}{Value of lambda used in the original Box-Cox transformation.}
}
\value{
A numeric vector or matrix with the original data restored (inverse transformation).
}
\description{
A function to compute the inverse Box-Cox transformation for a given transformed dataset \code{x}
and a given lambda value. The transformation returns the original pre-transformation data.
}
