% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_plot.R
\name{box_plot}
\alias{box_plot}
\title{Box-Cox Transformation Visualization}
\usage{
box_plot(x, transformed_x = NULL, lambda = NULL, plot_type = "histogram")
}
\arguments{
\item{x}{A numeric vector of the original data.}

\item{transformed_x}{An optional numeric vector of the transformed data. If not provided, the function will calculate it.}

\item{lambda}{Transformation parameter for the Box-Cox transformation (ignored if \code{transformed_x} is provided).}

\item{plot_type}{The type of plot to generate: "scatter" (default), histogram" or "density"}
}
\value{
A ggplot2 object displaying the comparison of original and transformed data.
}
\description{
This function compares the distribution of the original data with the transformed data
using side-by-side histograms or density plots.
}
\examples{
x <- c(1, 2, 3, 4, 10, 15, 25)
box_plot(x, lambda = 0.5)
}
\author{
Olivia Summerville
}
