% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimallambda.R
\name{estimate_optimal_lambda}
\alias{estimate_optimal_lambda}
\title{Estimate Optimal Lambda for Box-Cox Transformation}
\usage{
estimate_optimal_lambda(object, ...)
}
\arguments{
\item{object}{An object for which the optimal lambda is estimated.}

\item{...}{Additional arguements passed to specific methods}
}
\description{
This method estimates the optimal value of lambda for a Box-Cox transformation directly from a data frame or matrix.
The method calculates log-likelihood values for a range of lambda values and selects the optimal lambda.
}
